# Improve `coalesce` and `concat` performance for views