@alamb 2025-06-06
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                   coalesce_improvements                  main
-----                                                   ---------------------                  ----
concat 1024 arrays boolean 4                            1.00     28.7±0.04µs        ? ?/sec    1.01     28.9±0.05µs        ? ?/sec
concat 1024 arrays i32 4                                1.00     15.9±0.05µs        ? ?/sec    1.02     16.2±0.03µs        ? ?/sec
concat 1024 arrays str 4                                1.03     58.4±0.44µs        ? ?/sec    1.00     56.9±0.40µs        ? ?/sec
concat boolean 1024                                     1.00    437.8±1.03ns        ? ?/sec    1.04    457.4±2.22ns        ? ?/sec
concat boolean 8192 over 100 arrays                     1.00     50.9±0.08µs        ? ?/sec    1.00     51.0±0.06µs        ? ?/sec
concat boolean nulls 1024                               1.00    775.3±4.38ns        ? ?/sec    1.02    789.7±0.85ns        ? ?/sec
concat boolean nulls 8192 over 100 arrays               1.00    109.7±0.17µs        ? ?/sec    1.00    109.7±0.20µs        ? ?/sec
concat fixed size lists                                 1.00   714.2±17.89µs        ? ?/sec    1.01   721.3±17.25µs        ? ?/sec
concat i32 1024                                         1.01    446.1±0.77ns        ? ?/sec    1.00    439.6±0.93ns        ? ?/sec
concat i32 8192 over 100 arrays                         1.00   206.2±11.75µs        ? ?/sec    1.03    211.9±5.68µs        ? ?/sec
concat i32 nulls 1024                                   1.00    752.6±1.87ns        ? ?/sec    1.02    764.4±1.02ns        ? ?/sec
concat i32 nulls 8192 over 100 arrays                   1.00   286.0±11.26µs        ? ?/sec    1.03    294.2±8.00µs        ? ?/sec
concat str 1024                                         1.00     13.2±0.88µs        ? ?/sec    1.05     13.8±1.19µs        ? ?/sec
concat str 8192 over 100 arrays                         1.01    102.5±0.99ms        ? ?/sec    1.00    101.9±0.92ms        ? ?/sec
concat str nulls 1024                                   1.01      6.9±0.71µs        ? ?/sec    1.00      6.8±0.81µs        ? ?/sec
concat str nulls 8192 over 100 arrays                   1.00     52.3±0.42ms        ? ?/sec    1.00     52.1±0.47ms        ? ?/sec
concat str_dict 1024                                    1.00      2.8±0.01µs        ? ?/sec    1.00      2.8±0.02µs        ? ?/sec
concat str_dict_sparse 1024                             1.00      6.9±0.02µs        ? ?/sec    1.01      6.9±0.03µs        ? ?/sec
concat struct with int32 and dicts size=1024 count=2    1.00      6.4±0.15µs        ? ?/sec    1.03      6.6±0.18µs        ? ?/sec
```


</p>
</details>