@alamb 2025-06-06
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing coalesce_improvements (7b408ad051c9130410f40ddce7cc9fde922274a0) to f92ff183aad326b1a818f6495215f60b27d9ad4b [diff](https://github.com/apache/arrow-rs/compare/f92ff183aad326b1a818f6495215f60b27d9ad4b..7b408ad051c9130410f40ddce7cc9fde922274a0)
BENCH_NAME=filter_kernels
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench filter_kernels 
BENCH_FILTER=
BENCH_BRANCH_NAME=coalesce_improvements
Results will be posted here when complete