@alamb 2025-06-06
The filter improvement are 🌶️ 

```
group                                                            coalesce_improvements                  main
-----                                                            ---------------------                  ----
filter: mixed_utf8view, 8192, nulls: 0, selectivity: 0.001       1.00     60.5±0.53ms        ? ?/sec    1.30     78.5±0.94ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0, selectivity: 0.01        1.00      7.4±0.03ms        ? ?/sec    1.23      9.1±0.03ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0, selectivity: 0.1         1.00      3.2±0.05ms        ? ?/sec    1.31      4.2±0.16ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0, selectivity: 0.8         1.00      2.7±0.01ms        ? ?/sec    1.71      4.6±0.02ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0.1, selectivity: 0.001     1.00     79.4±0.47ms        ? ?/sec    1.19     94.4±1.03ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0.1, selectivity: 0.01      1.00     11.0±0.04ms        ? ?/sec    1.22     13.4±0.06ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0.1, selectivity: 0.1       1.00      4.5±0.07ms        ? ?/sec    1.36      6.2±0.17ms        ? ?/sec
filter: mixed_utf8view, 8192, nulls: 0.1, selectivity: 0.8       1.00      5.2±0.02ms        ? ?/sec    1.48      7.7±0.02ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0, selectivity: 0.001      1.00     85.9±0.32ms        ? ?/sec    1.49    127.6±0.56ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0, selectivity: 0.01       1.00     11.5±0.03ms        ? ?/sec    1.34     15.4±0.08ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0, selectivity: 0.1        1.00      5.3±0.23ms        ? ?/sec    1.42      7.5±0.15ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0, selectivity: 0.8        1.00      4.2±0.03ms        ? ?/sec    2.16      9.1±0.02ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.001    1.00    113.3±1.07ms        ? ?/sec    1.34    152.3±1.56ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.01     1.00     15.8±0.09ms        ? ?/sec    1.35     21.3±0.07ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.1      1.00      6.7±0.10ms        ? ?/sec    1.54     10.4±0.22ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.8      1.00      6.8±0.03ms        ? ?/sec    1.89     12.8±0.04ms        ? ?/sec
```